Lots of changes.

+++ mail-check.sh
@@ -4,11 +4,13 @@
 test -d "$XDG_DATA_HOME" || exit 0
 
 # Check for pid (remove stale).
-kill -0 $(cat $XDG_DATA_HOME/offlineimap/pid) 2>/dev/null && exit
-rm "$XDG_DATA_HOME/offlineimap/pid"
+if [[ -e "$XDG_DATA_HOME/offlineimap/pid" ]]; then
+    kill -0 $(cat "$XDG_DATA_HOME/offlineimap/pid") 2>/dev/null && exit
+    rm "$XDG_DATA_HOME/offlineimap/pid"
+fi
 
 # Restart on fail, die if killed.
-trap 'pkill -HUP -P $$ -x offlineimap; exit 0' 0
+trap 'pkill -USR2 -P $$ -x offlineimap; exit 0' 0
 while true; do
     # Wait until we have an internet connection.
     while ! ping -nc1 google.com 2>&1 >/dev/null; do
@@ -16,7 +18,7 @@ while true; do
     done
 
     mc=0
-    offlineimap -u Basic 2>/dev/null |
+    { ionice -c2 -n7 nice -n 10 offlineimap -u Basic 2>/dev/null |
     while read line; do
         if echo $line | grep -e "^Sleeping " >/dev/null; then
             if [[ $mc -gt 0 ]]; then
@@ -36,5 +38,5 @@ while true; do
         else
             echo "$line" | grep -e '^Copy message [0-9]* \w*\[INBOX\] ->.*$' > /dev/null && mc=$(( $mc + 1 ))
         fi
-    done
+    done } && exit 0
 done
